requirements.txt
anndata==0.8.0
black==22.6.0
boto3==1.24.43
botocore==1.27.43
click==8.1.3
cycler==0.11.0
fonttools==4.34.4
h5py==3.7.0
importlib-metadata==4.12.0
jmespath==1.0.1
joblib==1.1.0
kiwisolver==1.4.4
llvmlite==0.39.0
matplotlib==3.5.2
mypy-extensions==0.4.3
natsort==8.1.0
networkx==2.8.5
numba==0.56.0
numpy==1.22.4
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
patsy==0.5.2
Pillow==9.2.0
platformdirs==2.5.2
pyarrow==8.0.0
pynndescent==0.5.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
s3transfer==0.6.0
scanpy==1.9.1
scikit-learn==1.1.1
scipy==1.9.0
seaborn==0.11.2
session-info==1.0.0
six==1.16.0
statsmodels==0.13.2
stdlib-list==0.8.0
threadpoolctl==3.1.0
tiledb==0.17.1
tiledbsc==0.1.7
tomli==2.0.1
tqdm==4.64.0
typing_extensions==4.3.0
umap-learn==0.5.3
urllib3==1.26.11
zipp==3.8.1
